A-Package
Divide classes into packages
Enums, Parser, Ui, StoredList, Storage packaged as Utils
All Errors are packaged as ChatterBoxErrors
All types of Tasks are packaged as Tasks
