+++ src/main/java/ChatterBoxErrors/ChatterBoxDeadlineError.java
@@ -1,3 +1,5 @@
+package ChatterBoxErrors;
+
 public class ChatterBoxDeadlineError extends ChatterBoxError {
     public ChatterBoxDeadlineError() {
         super("____________________________________________________________\n"
