+++ src/main/java/ChatterBoxErrors/ChatterBoxError.java
@@ -1,3 +1,5 @@
+package ChatterBoxErrors;
+
 public class ChatterBoxError extends Exception {
     public ChatterBoxError() {
         super("____________________________________________________________\n"
