+++ src/main/java/ChatterBoxErrors/ChatterBoxMarkError.java
@@ -1,3 +1,5 @@
+package ChatterBoxErrors;
+
 public class ChatterBoxMarkError extends ChatterBoxError {
     public ChatterBoxMarkError() {
         super("____________________________________________________________\n"
