+++ src/main/java/Tasks/Deadline.java
@@ -1,35 +1,31 @@
-import java.time.LocalDate;
+package Tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task {
-    private LocalDateTime start;
+public class Deadline extends Task {
     private LocalDateTime end;
 
-    public Event (String name, LocalDateTime start, LocalDateTime end) {
+    public Deadline(String name, LocalDateTime by) {
         super(name);
-        this.start = start;
-        this.end = end;
+        this.end = by;
+
     }
 
     @Override
     public String storeTask() {
         StringBuilder saveTaskInfo = new StringBuilder();
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy HHmm");
-        String startDate = this.start.format(formatter);
-        String endDate = this.end.format(formatter);
         if (this.getCompleted()) {
             saveTaskInfo.append("done, ");
         }
         else {
             saveTaskInfo.append("undone, ");
         }
-        saveTaskInfo.append("event ");
+        saveTaskInfo.append("deadline ");
         saveTaskInfo.append(this.getName());
-        saveTaskInfo.append(" /from ");
-        saveTaskInfo.append(startDate);
-        saveTaskInfo.append(" /to ");
-        saveTaskInfo.append(endDate);
+        saveTaskInfo.append(" /by ");
+        saveTaskInfo.append(this.end.format(formatter));
         saveTaskInfo.append("\n");
         return saveTaskInfo.toString();
     }
@@ -37,8 +33,7 @@ public class Event extends Task {
     @Override
     public String toString() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd MMM yyyy HH:mm");
-        String startDate = this.start.format(formatter);
-        String endDate = this.end.format(formatter);
-        return "[E] " + super.toString() + " (from: " + startDate + " to: " + endDate + ")";
+        String date = this.end.format(formatter);
+        return "[D] " + super.toString() + " (by: " + date + ")";
     }
 }
