+++ src/main/java/Utils/Commands.java
@@ -1,0 +1,12 @@
+package Utils;
+
+public enum Commands {
+    BYE,
+    LIST,
+    MARK,
+    UNMARK,
+    DELETE,
+    TODO,
+    DEADLINE,
+    EVENT,
+}
