+++ src/main/java/Utils/Ui.java
@@ -1,3 +1,5 @@
+package Utils;
+
 import java.io.PrintStream;
 import java.util.Scanner;
